put filter in a separate repo
